rename reverse_hits.rs
